feat : 바이오리듬 커스텀 기반 영화 추천 API

+++ src/main/java/Recommend/Movie/Biorhythm/Dto/BiorhythmScore.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class BiorhythmScore {
-    int physical, emotional, intellectual;
-    public BiorhythmScore(int p, int e, int i) { this.physical = p; this.emotional = e; this.intellectual = i; }
+    double physical, emotional, intellectual;
+    public BiorhythmScore(double p, double e, double i) { this.physical = p; this.emotional = e; this.intellectual = i; }
 
 }
